--- crates/dwn-rs-messagestore/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "dwn-rs-messagestore"
-version = "0.2.0"
-edition = "2021"
-license = "Apache-2.0"
-description = "Message store implementations for dwn-rs"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[features]
-default = ["surreal"]
-surreal = ["surrealdb/rustls", "surrealdb/kv-mem", "surrealdb/kv-speedb", "surrealdb/protocol-ws", "surrealdb/protocol-http"]
-surreal-wasm = ["surrealdb/rustls", "surrealdb/kv-mem", "surrealdb/kv-indxdb", "surrealdb/protocol-ws", "surrealdb/protocol-http"]
-
-[dependencies]
-async-trait = "0.1.73"
-libipld = { version = "0.16.0", features = ["serde-codec"] }
-libipld-cbor = "0.16.0"
-libipld-core = { version = "0.16.0", features = [
-    "serde",
-    "serde-codec",
-    "multibase",
-] }
-serde = { version = "1.0.183", features = ["derive"] }
-surrealdb = { version = "1.1.1", default-features = false, optional = true }
-thiserror = "1.0.47"
-
-dwn-rs-core = { path = "../dwn-rs-core" }
-dwn-rs-stores = { path = "../dwn-rs-stores" }
