--- crates/dwn-rs-messagestore/src/lib.rs
@@ -1,5 +1,0 @@
-// #[cfg(feature = "surrealdb")]
-pub mod surrealdb;
-
-// #[cfg(feature = "surrealdb")]
-pub use surrealdb::*;
