+++ crates/dwn-rs-stores/src/lib.rs
@@ -3,5 +3,11 @@ pub mod filters;
 pub mod stores;
 
 pub use errors::*;
-pub use filters::*;
+pub use filters::{errors::*, filters::*, indexes::*, query::*, value::*};
 pub use stores::*;
+
+#[cfg(feature = "surreal")]
+pub mod surrealdb;
+
+#[cfg(feature = "surreal")]
+pub use surrealdb::*;
