+++ crates/dwn-rs-wasm/src/surrealdb/mod.rs
@@ -1,0 +1,2 @@
+pub mod data_store;
+pub mod message_store;
