+++ crates/dwn-rs-wasm/tests/test.js
@@ -1,10 +1,10 @@
 import { TestSuite } from "@tbd54566975/dwn-sdk-js/tests";
-import { SurrealDB } from "../pkg/index.js";
+import { SurrealMessageStore } from "../pkg/index.js";
 import WebSocket from "isomorphic-ws";
 
 global.WebSocket = WebSocket;
 
-let s = new SurrealDB();
+let s = new SurrealMessageStore();
 // await s.connect("ws://192.168.10.56:8000/");
 await s.connect("mem://");
 describe("Store dependent tests", () => {
